+++ src/main/java/br/com/schumaker/force/framework/jdbc/RdbEnum.java
@@ -1,4 +1,4 @@
-package br.com.schumaker.octopus.framework.jdbc;
+package br.com.schumaker.force.framework.jdbc;
 
 /**
  * The RdbEnum class.
